Revert "Features/mb.rezaee/add media entity/ in crud"

+++ Domain/Entities/BaseEntity.cs
@@ -1,10 +1,12 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Text;
 
 namespace Domain.Domain.Entities
 {
     public class BaseEntity
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public DateTime CreationDate { get; set; }
         public DateTime LastUpdateDate { get; set; }
     }
